feat: signature handler in order to centralize code
