+++ packages/network/src/signer/signers/vechain-base-signer/vechain-base-signer.ts
@@ -1,19 +1,10 @@
-import { type AvailableVechainProviders, type VechainSigner } from '../types';
 import {
-    DelegationHandler,
-    type SignTransactionOptions,
-    type ThorClient
-} from '../../../thor-client';
-import {
-    clauseBuilder,
-    Hex0x,
-    secp256k1,
-    Transaction,
-    type TransactionBody,
-    type TransactionClause,
-    TransactionHandler
-} from '../../../../../core';
-import { type TransactionObjectInput } from '../../../provider';
+    type AvailableVechainProviders,
+    type TransactionRequestInput,
+    type VechainSigner
+} from '../types';
+import { type SignTransactionOptions } from '../../../thor-client';
+import { signTransactionWithPrivateKey } from '../helpers';
 
 /**
  * Basic vechain signer.
@@ -35,7 +26,7 @@ class VechainBaseSigner<TProviderType extends AvailableVechainProviders>
      * @param provider - The provider to connect to
      */
     constructor(
-        readonly privateKey: Buffer,
+        private readonly privateKey: Buffer,
         provider: TProviderType | null
     ) {
         // Store provider and delegator
@@ -50,12 +41,13 @@ class VechainBaseSigner<TProviderType extends AvailableVechainProviders>
      * @returns The fully signed transaction
      */
     async signTransaction(
-        transactionToSign: TransactionObjectInput
+        transactionToSign: TransactionRequestInput
     ): Promise<string> {
-        return await this._sign(
+        return await signTransactionWithPrivateKey(
             transactionToSign,
             null,
-            (this.provider as TProviderType).thorClient
+            (this.provider as TProviderType).thorClient,
+            this.privateKey
         );
     }
 
@@ -67,128 +59,15 @@ class VechainBaseSigner<TProviderType extends AvailableVechainProviders>
      * @returns the fully signed transaction
      */
     async signTransactionWithDelegator(
-        transactionToSign: TransactionObjectInput,
+        transactionToSign: TransactionRequestInput,
         delegator: SignTransactionOptions
     ): Promise<string> {
-        return await this._sign(
+        return await signTransactionWithPrivateKey(
             transactionToSign,
             delegator,
-            (this.provider as TProviderType).thorClient
-        );
-    }
-
-    /**
-     * Signs a transaction internal method
-     */
-    private async _sign(
-        transaction: TransactionObjectInput,
-        delegator: SignTransactionOptions | null,
-        thorClient: ThorClient
-    ): Promise<string> {
-        // 1 - Initiate the transaction clauses
-        const transactionClauses: TransactionClause[] =
-            transaction.to !== undefined
-                ? // Normal transaction
-                  [
-                      {
-                          to: transaction.to,
-                          data: transaction.data ?? '0x',
-                          value: transaction.value ?? '0x0'
-                      } satisfies TransactionClause
-                  ]
-                : // If 'to' address is not provided, it will be assumed that the transaction is a contract creation transaction.
-                  [clauseBuilder.deployContract(transaction.data ?? '0x')];
-
-        // 2 - Estimate gas
-        const gasResult = await thorClient.gas.estimateGas(
-            transactionClauses,
-            transaction.from
-        );
-
-        // 3 - Create transaction body
-        const transactionBody =
-            await thorClient.transactions.buildTransactionBody(
-                transactionClauses,
-                gasResult.totalGas,
-                {
-                    isDelegated: DelegationHandler(delegator).isDelegated()
-                }
-            );
-
-        // 5 - Generate nonce.
-        // @NOTE: To be compliant with the standard and to avoid nonce overflow, we generate a random nonce of 6 bytes
-
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const { nonce, ...transactionBodyWithoutNonce } = transactionBody;
-        const newNonce = Hex0x.of(secp256k1.randomBytes(6));
-
-        const finalTransactionBody: TransactionBody = {
-            nonce: newNonce,
-            ...transactionBodyWithoutNonce
-        };
-
-        // 6 - Sign the transaction
-
-        const signedTransaction = DelegationHandler(delegator).isDelegated()
-            ? await this._signTransactionBodyWithDelegator(
-                  transaction.from,
-                  finalTransactionBody,
-                  delegator as SignTransactionOptions,
-                  thorClient
-              )
-            : TransactionHandler.sign(finalTransactionBody, this.privateKey);
-
-        return Hex0x.of(signedTransaction.encoded);
-    }
-
-    /**
-     * Sign a transaction with the delegator.
-     * The signature of the delegator into the wallet will be used to sign the transaction.
-     *
-     * @param transactionOrigin - The origin address of the transaction (the 'from' field).
-     * @param transactionToSign - The transaction to sign.
-     * @param thorClient - The ThorClient instance used to sign using the url
-     * @returns The transaction signed by the delegator.
-     *
-     */
-    private async _signTransactionBodyWithDelegator(
-        transactionOrigin: string,
-        transactionToSign: TransactionBody,
-        delegator: SignTransactionOptions,
-        thorClient: ThorClient
-    ): Promise<Transaction> {
-        // 1 - Sign with delegatorPrivateKey
-        if (delegator?.delegatorPrivateKey !== undefined) {
-            return TransactionHandler.signWithDelegator(
-                transactionToSign,
-                this.privateKey,
-                Buffer.from(delegator.delegatorPrivateKey, 'hex')
-            );
-        }
-
-        // 2 - Sign the transaction with the delegator url
-        const delegatorSignatureWithUrl = await DelegationHandler(
-            delegator
-        ).getDelegationSignatureUsingUrl(
-            new Transaction(transactionToSign),
-            transactionOrigin,
-            thorClient.httpClient
-        );
-
-        // Sign transaction with origin private key
-        const originSignature = secp256k1.sign(
-            new Transaction(transactionToSign).getSignatureHash(),
+            (this.provider as TProviderType).thorClient,
             this.privateKey
         );
-
-        // Sign the transaction with both signatures. Concat both signatures to get the final signature
-        const signature = Buffer.concat([
-            originSignature,
-            delegatorSignatureWithUrl
-        ]);
-
-        // Return new signed transaction
-        return new Transaction(transactionToSign, signature);
     }
 }
